Terminando segunda parte

+++ database/migrations/2023_03_29_111047_create_pagamentos_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('banco');
             $table->string('anexo');
             $table->enum('estado',['validado','invalidado','pendente'])->default('pendente');
-            $table->integer('quantia');
+            $table->decimal('quantia', 10, 2);
             $table->timestamps();
 
             $table->foreign('user_id')
